Add example of using fs.readFileSync for newbies
The removal of `fromFile` is confusing for those who find this plugin from npmjs.org. That is because the old version of the README is still on npmjs.org dispite an updated version from `npm install`.

Others like me might be confused. I was smart enough to search github for issue #4 and was lucky to see someone commented the use of `fs.readFileSync`.

This change adds an example to the README for others who might be interested in the reading of a file but are not sure or new to node to realize that `fileReadSync` works in this case.